Big Update!
In the sketches now the communication is at 115200 baud so it is faster
All the edits are in the receiver sketch:
- Now we don't work anymore with data types uintX_t but only with
integers
- The convertIR function has been incorporated by the listenForIR
function that now while receiving alerady detects if the code that is
arriving is good or not and what the pulse received corresponds to. So
all the data goes in a data[] array that holds everything. Easier, isn't
it?